fix: refactor to pull item parsing into its own dir
