+++ src/module/actors.convert.ts
@@ -1,14 +1,13 @@
-import { Abilities, ArmorClass, Feature, Health, ImportActor, parseFormula, Skill } from './actors.process';
+import { Abilities, ArmorClass, Feature, Health, ImportActor, Skill } from './actors.process';
 import {
   FifthAbilities,
   FifthAttributes,
-  FifthFeatureCost,
   FifthItem,
-  FifthItemType,
   FifthSkill,
   FifthSkills,
   FifthStat,
 } from './fifthedition.actor.template';
+import { featuresToItems } from './item/weapon';
 
 export function convertAbilities({ str, dex, con, int, wis, cha }: Abilities): FifthAbilities {
   return {
@@ -136,30 +135,6 @@ function convertSkills(skills: Skill[]): FifthSkills {
   return fifthSkills;
 }
 
-function getDamageType(from: string): string | undefined {
-  if (from.includes('piercing')) return 'piercing';
-  if (from.includes('slashing')) return 'slashing';
-  if (from.includes('bludgeoning')) return 'bludgeoning';
-  if (from.includes('fire')) return 'fire';
-  if (from.includes('cold')) return 'cold';
-  if (from.includes('lightning')) return 'lightning';
-  if (from.includes('acid')) return 'acid';
-  if (from.includes('poison')) return 'poison';
-  if (from.includes('psychic')) return 'psychic';
-  if (from.includes('radiant')) return 'radiant';
-  if (from.includes('thunder')) return 'thunder';
-  if (from.includes('force')) return 'force';
-  if (from.includes('necrotic')) return 'necrotic';
-  if (from.includes('psychic')) return 'psychic';
-}
-
-export function buildDamageParts(description: string) {
-  // description = 'Melee Weapon Attack: +6 to hit, reach 5 ft., one target.Hit: 8 (1d8 + 4) piercing damage.'
-  const parsed = parseFormula(description, /Melee Weapon Attack: +/);
-  const fromString = parsed.afterFormula ? parsed.afterFormula : description;
-  return [[parsed.str, getDamageType(fromString)]];
-}
-
 export function buildAttackBonus(description: string, mod: number): number | undefined {
   // regex for + followed by one or two digits
   const attackMatch = description.match(/\+\d{1,2}/);
@@ -192,53 +167,6 @@ export function buildReach(description: string) {
     value: Number(value),
     units: type,
   };
-}
-
-function getActionType(description: string): string | undefined {
-  if (/melee/i.test(description)) return 'mwak';
-  if (/ranged/.test(description)) return 'rwak';
-  return undefined;
-}
-
-function getItemType(description: string): FifthItemType {
-  let itemType: FifthItemType = 'feat';
-  if (/melee weapon attack/i.test(description)) itemType = 'weapon';
-  if (/ranged weapon attack/i.test(description)) itemType = 'weapon';
-  if (/melee or ranged weapon attack/i.test(description)) itemType = 'weapon';
-  return itemType;
-}
-
-function getMaxAbility(abilities: Abilities): FifthStat {
-  if (abilities.str.mod > abilities.dex.mod) return 'str';
-  return 'dex';
-}
-
-export function featuresToItems(features: Feature[], abilities: Abilities): FifthItem[] {
-  return features.map((feature) => {
-    let activationType: FifthFeatureCost = 'none';
-    const itemType: FifthItemType = getItemType(feature.description);
-    if (feature.description.includes('action')) activationType = 'action';
-    if (feature.description.includes('bonus action')) activationType = 'bonus';
-
-    const damage = itemType === 'weapon' ? { parts: buildDamageParts(feature.description) } : {};
-    const ability = getMaxAbility(abilities);
-
-    return {
-      name: feature.name,
-      type: itemType,
-      data: {
-        description: {
-          value: feature.description,
-        },
-        activation: {
-          type: activationType,
-        },
-        damage,
-        actionType: getActionType(feature.description),
-        ability,
-      },
-    };
-  });
 }
 
 export interface FeatureCollection {
