+++ test/actors.convert.test.ts
@@ -1,4 +1,4 @@
-import { actorToFifth, buildAttackBonus, buildDamageParts, buildReach } from '../src/module/actors.convert';
+import { actorToFifth, buildAttackBonus, buildReach } from '../src/module/actors.convert';
 import { swashbuckler } from './mock.swashbuckler';
 
 describe('Parsed Actor to 5th Structure', () => {
@@ -8,20 +8,6 @@ describe('Parsed Actor to 5th Structure', () => {
     expect(fifthSwashbuckler.attributes.ac.flat).toEqual(17);
     if (!fifthSwashbuckler.skills.acr) throw new Error('Missing ACR');
     expect(fifthSwashbuckler.skills.acr.value).toEqual(8);
-  });
-});
-
-describe('buildDamage', () => {
-  it('should build damage from a melee weapon', () => {
-    const rapier = 'Melee Weapon Attack: +6 to hit, reach 5 ft., one target.Hit: 8 (1d8 + 4) piercing damage.';
-    const built = buildDamageParts(rapier);
-    expect(built).toEqual([['1d8 + 4', 'piercing']]);
-  });
-
-  it('should build damage for a bite attack', () => {
-    const bite = 'Bite. Melee Weapon Attack: +5 to hit, reach 5 ft., one target. Hit: 8 (2d4 + 3) piercing damage.';
-    const built = buildDamageParts(bite);
-    expect(built).toEqual([['2d4 + 3', 'piercing']]);
   });
 });
